requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
colorama==0.4.3
cryptography==3.1.1
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.61.1
h11==0.10.0
idna==2.10
iniconfig==1.0.1
packaging==20.4
passlib==1.7.2
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.1.0
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
toml==0.10.1
uvicorn==0.12.1
